Combine two independent tests for thread stress/cleanup.
