--- src/test/thread_cleanup.run
@@ -1,2 +1,0 @@
-source `dirname $0`/util.sh thread_cleanup "$@"
-compare_test EXIT-SUCCESS
