+++ src/test/thread_stress.run
@@ -1,0 +1,2 @@
+source `dirname $0`/util.sh thread_stress "$@"
+compare_test EXIT-SUCCESS
